+++ packages/stream_feed/CHANGELOG.md
@@ -3,10 +3,10 @@
 - fix(serverside llc): `issueJwtHS256` wasn't using the `expiresAt` field and remove the default "exp" value.
 - new: pagination suuport for flat feed activities. For example:
 ```dart
-final paginated = await flatFeed.paginatedActivities();
+final paginated = await flatFeed.getPaginatedEnrichedActivities();
 final nextParams = parseNext(paginated.next!);
 // parse next page
-await flatFeed.paginatedActivities(limit: nextParams.limit,filter: nextParams.idLT);
+await flatFeed.getPaginatedEnrichedActivities(limit: nextParams.limit,filter: nextParams.idLT);
 ```
 
 ## 0.5.1+1: 24/03/2022